+++ tests/__main__.py
@@ -13,6 +13,7 @@ from .test_powertrain import main as powertrain_main
 from .test_surrogates import main as surrogates_main
 from .test_data_model import main as lstm_main
 from .test_direct import main as direct_main
+from .test_linear_model import main as linear_main
 
 from io import StringIO
 import matplotlib.pyplot as plt
@@ -109,6 +110,11 @@ if __name__ == '__main__':
 
     try:
         lstm_main()
+    except Exception:
+        was_successful = False
+
+    try:
+        linear_main()
     except Exception:
         was_successful = False
